Add `app-k9mail` module to host the app project
