+++ app-k9mail/dependencies/releaseRuntimeClasspath.txt
