+++ app-k9mail/src/debug/java/app/k9mail/dev/DebugConfig.kt
@@ -1,0 +1,12 @@
+package app.k9mail.dev
+
+import com.fsck.k9.backend.BackendFactory
+import org.koin.core.module.Module
+import org.koin.core.qualifier.named
+
+fun Module.developmentModuleAdditions() {
+    single { DemoBackendFactory(backendStorageFactory = get()) }
+    single<Map<String, BackendFactory>>(named("developmentBackends")) {
+        mapOf("demo" to get<DemoBackendFactory>())
+    }
+}
