+++ app-k9mail/src/debug/res/values/app_logo_colors.xml
