+++ app-k9mail/src/main/AndroidManifest.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    android:installLocation="auto">
+
+    <application
+        android:name="com.fsck.k9.K9App"
+        android:icon="@drawable/ic_launcher"
+        android:label="@string/app_name" />
+</manifest>
