+++ app-k9mail/src/main/kotlin/com/fsck/k9/K9App.kt
@@ -1,0 +1,7 @@
+package com.fsck.k9
+
+import org.koin.core.module.Module
+
+class K9App : CommonApp() {
+    override fun provideAppModule(): Module = appModule
+}
