+++ app-k9mail/src/main/res/values/drawables.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <drawable name="ic_launcher">@drawable/ic_app_logo</drawable>
+</resources>
