+++ app-k9mail/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
+    <!-- Used in AndroidManifest.xml -->
+    <string name="app_name">K-9 Mail</string>
+</resources>
