+++ app/k9mail/build.gradle.kts
@@ -1,11 +1,5 @@
 plugins {
-    id(ThunderbirdPlugins.App.android)
-    alias(libs.plugins.dependency.guard)
-}
-
-val testCoverageEnabled: Boolean by extra
-if (testCoverageEnabled) {
-    apply(plugin = "jacoco")
+    id(ThunderbirdPlugins.Library.android)
 }
 
 dependencies {
@@ -16,7 +10,6 @@ dependencies {
     implementation(projects.app.cryptoOpenpgp)
     implementation(projects.backend.imap)
     implementation(projects.backend.pop3)
-    debugImplementation(projects.backend.demo)
 
     implementation(projects.core.featureflags)
     implementation(projects.feature.launcher)
@@ -45,127 +38,14 @@ dependencies {
 }
 
 android {
-    namespace = "com.fsck.k9"
-
-    defaultConfig {
-        applicationId = "com.fsck.k9"
-        testApplicationId = "com.fsck.k9.tests"
-
-        versionCode = 37014
-        versionName = "6.715-SNAPSHOT"
-
-        // Keep in sync with the resource string array "supported_languages"
-        resourceConfigurations.addAll(
-            listOf(
-                "in", "br", "ca", "cs", "cy", "da", "de", "et", "en", "en_GB", "es", "eo", "eu", "fr", "gd", "gl",
-                "hr", "is", "it", "lv", "lt", "hu", "nl", "nb", "pl", "pt_PT", "pt_BR", "ru", "ro", "sq", "sk", "sl",
-                "fi", "sv", "tr", "el", "be", "bg", "sr", "uk", "iw", "ar", "fa", "ml", "ko", "zh_CN", "zh_TW", "ja",
-                "fy",
-            ),
-        )
-
-        buildConfigField("String", "CLIENT_ID_APP_NAME", "\"K-9 Mail\"")
-    }
-
-    signingConfigs {
-        if (project.hasProperty("k9mail.keyAlias") &&
-            project.hasProperty("k9mail.keyPassword") &&
-            project.hasProperty("k9mail.storeFile") &&
-            project.hasProperty("k9mail.storePassword")
-        ) {
-            create("release") {
-                keyAlias = project.property("k9mail.keyAlias") as String
-                keyPassword = project.property("k9mail.keyPassword") as String
-                storeFile = file(project.property("k9mail.storeFile") as String)
-                storePassword = project.property("k9mail.storePassword") as String
-            }
-        }
-    }
+    namespace = "com.fsck.k9.common"
 
     buildTypes {
-        release {
-            signingConfig = signingConfigs.findByName("release")
-
-            isMinifyEnabled = true
-            proguardFiles(
-                getDefaultProguardFile("proguard-android.txt"),
-                "proguard-rules.pro",
-            )
-
-            buildConfigField(
-                "String",
-                "OAUTH_GMAIL_CLIENT_ID",
-                "\"262622259280-hhmh92rhklkg2k1tjil69epo0o9a12jm.apps.googleusercontent.com\"",
-            )
-            buildConfigField(
-                "String",
-                "OAUTH_YAHOO_CLIENT_ID",
-                "\"dj0yJmk9aHNUb3d2MW5TQnpRJmQ9WVdrOWVYbHpaRWM0YkdnbWNHbzlNQT09JnM9Y29uc3VtZXJzZWNyZXQmc3Y9MCZ4PWIz\"",
-            )
-            buildConfigField(
-                "String",
-                "OAUTH_AOL_CLIENT_ID",
-                "\"dj0yJmk9dUNqYXZhYWxOYkdRJmQ9WVdrOU1YQnZVRFZoY1ZrbWNHbzlNQT09JnM9Y29uc3VtZXJzZWNyZXQmc3Y9MCZ4PWIw\"",
-            )
-            buildConfigField("String", "OAUTH_MICROSOFT_CLIENT_ID", "\"e647013a-ada4-4114-b419-e43d250f99c5\"")
-            buildConfigField(
-                "String",
-                "OAUTH_MICROSOFT_REDIRECT_URI",
-                "\"msauth://com.fsck.k9/Dx8yUsuhyU3dYYba1aA16Wxu5eM%3D\"",
-            )
-
-            manifestPlaceholders["appAuthRedirectScheme"] = "com.fsck.k9"
-        }
-
         debug {
-            applicationIdSuffix = ".debug"
-            enableUnitTestCoverage = testCoverageEnabled
-            enableAndroidTestCoverage = testCoverageEnabled
-
-            isMinifyEnabled = false
-
-            buildConfigField(
-                "String",
-                "OAUTH_GMAIL_CLIENT_ID",
-                "\"262622259280-5qb3vtj68d5dtudmaif4g9vd3cpar8r3.apps.googleusercontent.com\"",
-            )
-            buildConfigField(
-                "String",
-                "OAUTH_YAHOO_CLIENT_ID",
-                "\"dj0yJmk9ejRCRU1ybmZjQlVBJmQ9WVdrOVVrZEViak4xYmxZbWNHbzlNQT09JnM9Y29uc3VtZXJzZWNyZXQmc3Y9MCZ4PTZj\"",
-            )
-            buildConfigField(
-                "String",
-                "OAUTH_AOL_CLIENT_ID",
-                "\"dj0yJmk9cHYydkJkTUxHcXlYJmQ9WVdrOWVHZHhVVXN4VVV3bWNHbzlNQT09JnM9Y29uc3VtZXJzZWNyZXQmc3Y9MCZ4PTdm\"",
-            )
-            buildConfigField("String", "OAUTH_MICROSOFT_CLIENT_ID", "\"e647013a-ada4-4114-b419-e43d250f99c5\"")
-            buildConfigField(
-                "String",
-                "OAUTH_MICROSOFT_REDIRECT_URI",
-                "\"msauth://com.fsck.k9.debug/VZF2DYuLYAu4TurFd6usQB2JPts%3D\"",
-            )
-
-            manifestPlaceholders["appAuthRedirectScheme"] = "com.fsck.k9.debug"
+            manifestPlaceholders["appAuthRedirectScheme"] = "FIXME: override this in your app project"
+        }
+        release {
+            manifestPlaceholders["appAuthRedirectScheme"] = "FIXME: override this in your app project"
         }
     }
-
-    packaging {
-        jniLibs {
-            excludes += listOf("kotlin/**")
-        }
-
-        resources {
-            excludes += listOf(
-                "META-INF/*.kotlin_module",
-                "META-INF/*.version",
-                "kotlin/**",
-                "DebugProbesKt.bin",
-            )
-        }
-    }
-}
-
-dependencyGuard {
-    configuration("releaseRuntimeClasspath")
 }
