+++ app/k9mail/src/main/java/com/fsck/k9/backends/RealOAuth2TokenProvider.kt
@@ -1,7 +1,6 @@
 package com.fsck.k9.backends
 
 import android.content.Context
-import com.fsck.k9.BuildConfig
 import com.fsck.k9.mail.AuthenticationFailedException
 import com.fsck.k9.mail.oauth.AuthStateStorage
 import com.fsck.k9.mail.oauth.OAuth2TokenProvider
@@ -18,6 +17,7 @@ import timber.log.Timber
 class RealOAuth2TokenProvider(
     context: Context,
     private val authStateStorage: AuthStateStorage,
+
 ) : OAuth2TokenProvider {
     private val authService = AuthorizationService(context)
     private var requestFreshToken = false
@@ -49,11 +49,6 @@ class RealOAuth2TokenProvider(
 
             latch.await(timeoutMillis, TimeUnit.MILLISECONDS)
         } catch (e: Exception) {
-            // OAuth errors are communicated via the callback. If we end up here, it's probably a programming error.
-            if (BuildConfig.DEBUG) {
-                throw AssertionError("Wrong usage of AuthState.performActionWithFreshTokens()?", e)
-            }
-
             Timber.w(e, "Failed to fetch an access token. Clearing authorization state.")
 
             authStateStorage.updateAuthorizationState(authorizationState = null)
