--- app/k9mail/src/test/java/com/fsck/k9/AppRobolectricTest.kt
@@ -1,14 +1,0 @@
-package com.fsck.k9
-
-import android.app.Application
-import org.junit.runner.RunWith
-import org.koin.test.AutoCloseKoinTest
-import org.robolectric.RobolectricTestRunner
-import org.robolectric.annotation.Config
-
-/**
- * A Robolectric test that creates an instance of our [Application] class [App].
- */
-@RunWith(RobolectricTestRunner::class)
-@Config(application = App::class)
-abstract class AppRobolectricTest : AutoCloseKoinTest()
